# be/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Clinic Management API",
        "title": "Clinic Management",
        "contact": {},
        "version": "1.0"
    },
    "host": "clinicly.fly.dev",
    "basePath": "/api/v1",
    "paths": {
...
        "/medicine": {
            "get": {
                "description": "Get medicine",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Get medicine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Medicine name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Order by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Order descending",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineListResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create medicine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Create medicine",
                "parameters": [
                    {
                        "description": "Medicine data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineResponse"
                        }
                    }
                }
            }
        },
...
        "/medicine/{id}": {
            "get": {
                "description": "Get medicine by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Get medicine by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update medicine",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Update medicine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medicine data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateMedicineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete medicine",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medicine"
                ],
                "summary": "Delete medicine",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medicine id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Medicine response",
                        "schema": {
                            "$ref": "#/definitions/controllers.MedicineResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.MedicineListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Medicine"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.MedicineRequest": {
            "type": "object",
            "required": [
                "id",
                "name",
                "price",
                "quantity",
                "unit"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "unit": {
                    "$ref": "#/definitions/types.MedicineUnit"
                }
            }
        },
        "controllers.MedicineResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Medicine"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.UpdateMedicineRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "unit": {
                    "$ref": "#/definitions/types.MedicineUnit"
                }
            }
        },
...
        "models.Medicine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "unit": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
        "types.MedicineUnit": {
            "type": "integer",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "Pill",
                "Bottle"
            ]
        },
...
    }
}